replace eventBuffer with coordMappings; compute tlbr-screen in browser.py
