--- experiments/dtw.py
@@ -1,412 +1,0 @@
-from statistics import mean, median, stdev
-
-from bs4 import BeautifulSoup
-from copy import deepcopy
-from dtaidistance import dtw, dtw_ndim
-from loguru import logger
-import numpy as np
-
-from openadapt import utils
-from openadapt.db import crud
-
-# action to browser
-MOUSE_BUTTON_MAPPING = {
-    "left": 0,
-    "right": 2,
-    "middle": 1
-}
-
-# action to browser
-EVENT_TYPE_MAPPING = {
-    "click": "click",
-    "press": "keydown",
-    "release": "keyup"
-}
-
-# TODO: read from pynput
-KEYBOARD_KEYS = [
-    'a', 'b', 'c', 'd', 'e', 'f', 'g', 'h', 'i', 'j', 'k', 'l', 'm', 'n', 'o', 'p',
-    'q', 'r', 's', 't', 'u', 'v', 'w', 'x', 'y', 'z',
-    '0', '1', '2', '3', '4', '5', '6', '7', '8', '9',
-    'tab', 'enter', 'shift', 'ctrl', 'alt', 'esc', 'space', 'backspace',
-    'delete', 'home', 'end', 'pageup', 'pagedown',
-    'arrowup', 'arrowdown', 'arrowleft', 'arrowright'
-]
-
-
-def main() -> None:
-    session = crud.get_new_session(read_only=True)
-    recording = crud.get_latest_recording(session)
-    action_events = crud.get_action_events(session=session, recording=recording)
-    browser_events = crud.get_browser_events(session=session, recording=recording)
-
-    # Filter BrowserEvents for 'USER_EVENT' type
-    browser_events = [
-        browser_event
-        for browser_event in browser_events
-        if browser_event.message["type"] == "USER_EVENT"
-    ]
-
-    # Define event pairs dynamically for mouse events
-    event_pairs = [
-        ("Left Click", "click", "left"),
-        ("Right Click", "click", "right"),
-        ("Middle Click", "click", "middle"),
-    ]
-
-    # Add keyboard events dynamically
-    event_pairs.extend(
-        [(f"Key Press {key.upper()}", "press", key) for key in KEYBOARD_KEYS] +
-        [(f"Key Release {key.upper()}", "release", key) for key in KEYBOARD_KEYS]
-    )
-
-    total_errors = 0  # Initialize total errors counter
-    total_remote_time_differences = []
-    total_local_time_differences = []
-    total_mouse_x_differences = []
-    total_mouse_y_differences = []
-
-    # Process each event pair
-    for event_type, action_name, key_or_button in event_pairs:
-        action_filtered_events = list(filter(lambda e: is_action_event(e, action_name, key_or_button), action_events))
-        browser_filtered_events = list(filter(lambda e: is_browser_event(e, action_name, key_or_button), browser_events))
-
-        # Only log if there are any action or browser events
-        if action_filtered_events or browser_filtered_events:
-            logger.info(f"{event_type}: {len(action_filtered_events)} action events, {len(browser_filtered_events)} browser events")
-
-            # Align and print events for the specific type
-            errors, remote_time_differences, local_time_differences, mouse_x_differences, mouse_y_differences = align_and_print(
-                event_type, action_filtered_events, browser_filtered_events
-            )
-            total_errors += errors  # Accumulate errors
-            total_remote_time_differences += remote_time_differences
-            total_local_time_differences += local_time_differences
-            total_mouse_x_differences += mouse_x_differences
-            total_mouse_y_differences += mouse_y_differences
-
-            # Identify and log the smallest clicked DOM element
-            for browser_event in browser_filtered_events:
-                if browser_event.message["eventType"] == "click":
-                    identify_and_log_smallest_clicked_element(browser_event)
-
-    # Calculate and log statistics for timestamp differences
-    for name, time_differences in (
-        ('Remote', total_remote_time_differences),
-        ('Local', total_local_time_differences),
-    ):
-        if not time_differences:
-            logger.warning(f"{name=} {time_differences=}")
-            continue
-        min_diff = min(time_differences, key=abs)
-        max_diff = max(time_differences, key=abs)
-        mean_diff = mean(time_differences)
-        median_diff = median(time_differences)
-        stddev_diff = stdev(time_differences) if len(time_differences) > 1 else 0
-        logger.info(
-            f"{name} Timestamp Differences - Min: {min_diff:.4f}, Max: {max_diff:.4f}, "
-            f"Mean: {mean_diff:.4f}, Median: {median_diff:.4f}, Std Dev: {stddev_diff:.4f}"
-        )
-
-    # Calculate and log statistics for mouse position differences
-    for axis, differences in (("X", total_mouse_x_differences), ("Y", total_mouse_y_differences)):
-        if not differences:
-            logger.warning(f"{axis=} {differences=}")
-            continue
-        min_diff = min(differences, key=abs)
-        max_diff = max(differences, key=abs)
-        num_errors = sum([diff >= 1 for diff in differences])
-        num_correct = sum([diff < 1 for diff in differences])
-        logger.info(f"{num_errors=} {num_correct=}")
-        if abs(max_diff) >= 1:
-            logger.warning(f"{abs(max_diff)=} > 1")
-        mean_diff = mean(differences)
-        median_diff = median(differences)
-        stddev_diff = stdev(differences) if len(differences) > 1 else 0
-        logger.info(
-            f"Mouse {axis} Position Differences - Min: {min_diff:.4f}, Max: {max_diff:.4f}, "
-            f"Mean: {mean_diff:.4f}, Median: {median_diff:.4f}, Std Dev: {stddev_diff:.4f}"
-        )
-
-    # Report total errors at the end
-    logger.info(f"Total Errors Across All Events: {total_errors}")
-
-def identify_and_log_smallest_clicked_element(browser_event) -> None:
-    """
-    Identifies the smallest DOM element that was clicked on for a given click event 
-    and logs it.
-
-    Args:
-        browser_event: The browser event containing the click details.
-    """
-    visible_html_string = browser_event.message.get('visibleHtmlString')
-    message_id = browser_event.message.get('id')
-    logger.info("*" * 10)
-    logger.info(f"{message_id=}")
-    target_id = browser_event.message.get('targetId')
-    logger.info(f"{target_id=}")
-
-    if not visible_html_string:
-        logger.warning("No visible HTML data available for click event.")
-        return
-
-    # Parse the visible HTML using BeautifulSoup
-    soup = BeautifulSoup(visible_html_string, 'html.parser')
-    target_element = soup.find(attrs={"data-id": target_id})
-    if not target_element:
-        logger.warning(f"{target_element=}")
-    else:
-        for coord_type in ('client', 'screen'):
-            x = browser_event.message.get(f"{coord_type}X")
-            y = browser_event.message.get(f"{coord_type}Y")
-            tlbr = f"data-tlbr-{coord_type}"
-            target_element_tlbr = target_element[tlbr]
-            top, left, bottom, right = map(float, target_element_tlbr.split(','))
-            logger.info(f"{tlbr=} {x=} {y=} {top=} {left=} {bottom=} {right=}")
-            if not (left <= x <= right and top <= y <= bottom):
-                #import ipdb; ipdb.set_trace()
-                logger.warning(f"outside")
-
-    elements = soup.find_all(attrs={'data-tlbr-client': True})
-    
-    smallest_element = None
-    smallest_area = float('inf')
-
-    for elem in elements:
-        #if elem['data-id'] == target_id:
-        #    import ipdb; ipdb.set_trace()
-        data_tlbr = elem['data-tlbr-client']
-        top, left, bottom, right = map(float, data_tlbr.split(','))
-        client_x = browser_event.message.get("clientX")
-        client_y = browser_event.message.get("clientY")
-
-        if left <= client_x <= right and top <= client_y <= bottom:
-            area = (right - left) * (bottom - top)
-            if area < smallest_area:
-                smallest_area = area
-                smallest_element = elem
-
-    logger.info(f"browser_event={str(browser_event)}")
-    if smallest_element is not None:
-        smallest_element_str = utils.truncate_html(str(smallest_element), 100)
-        logger.info(f"Smallest clicked element found: {tlbr} {smallest_element_str}")
-        smallest_element_id = smallest_element["data-id"]
-        assert smallest_element_id == target_id, (smallest_element_id, target_id)
-    else:
-        logger.warning("No element found matching the click coordinates.")
-
-def is_action_event(event, action_name: str, key_or_button: str) -> bool:
-    """
-    Determine if the event matches the given action name and key/button.
-
-    Args:
-        event: The action event to check.
-        action_name: The type of action (e.g., "click", "press", "release").
-        key_or_button: The key or button associated with the action.
-
-    Returns:
-        bool: True if the event matches the action name and key/button, False otherwise.
-    """
-    if action_name == "click":
-        return event.name == action_name and event.mouse_button_name == key_or_button
-    elif action_name in {"press", "release"}:
-        stripped_action_text = event._text(name_prefix="", name_suffix="")
-        return event.name == action_name and stripped_action_text == key_or_button
-    else:
-        return False
-
-
-def is_browser_event(event, action_name: str, key_or_button: str) -> bool:
-    if action_name == "click":
-        return (event.message.get("eventType") == EVENT_TYPE_MAPPING[action_name] and
-                event.message.get("button") == MOUSE_BUTTON_MAPPING[key_or_button])
-    elif action_name in {"press", "release"}:
-        return (event.message.get("eventType") == EVENT_TYPE_MAPPING[action_name] and
-                event.message.get("key").lower() == key_or_button)
-    else:
-        return False
-
-
-def composite_distance(
-    action_event, browser_event, 
-    time_weight: float = 1.0, 
-    spatial_weight: float = 1.0
-) -> float:
-    """
-    Computes a composite distance between action and browser events 
-    based on both time and spatial (x/y) position differences.
-
-    Args:
-        action_event: The action event containing timestamp and position data.
-        browser_event: The browser event containing timestamp and position data.
-        time_weight: The weight assigned to the time difference in the composite distance.
-        spatial_weight: The weight assigned to the spatial difference in the composite distance.
-
-    Returns:
-        float: The computed composite distance.
-    """
-    # Time difference
-    time_diff = abs(action_event.timestamp - browser_event.message["timestamp"])
-
-    # Spatial differences (x/y positions)
-    if action_event.mouse_x is not None and action_event.mouse_y is not None:
-        x_diff = action_event.mouse_x - browser_event.message.get('screenX', 0)
-        y_diff = action_event.mouse_y - browser_event.message.get('screenY', 0)
-        spatial_diff = (x_diff ** 2 + y_diff ** 2) ** 0.5  # Euclidean distance
-    else:
-        spatial_diff = 0
-
-    # Composite distance: weighted sum of time and spatial differences
-    return (time_weight * time_diff) + (spatial_weight * spatial_diff)
-
-
-def align_and_print(
-    event_type: str,
-    action_events: list,
-    browser_events: list,
-) -> tuple[int, list[float], list[float], list[float], list[float]]:
-    # Only log if there are any action or browser events of each type
-    if not action_events and not browser_events:
-        return 0, [], [], [], []
-
-    # Convert series of events to timestamps
-    action_timestamps = [e.timestamp for e in action_events]
-    #browser_timestamps = [e.message["timestamp"] for e in browser_filtered_events]
-    browser_timestamps = [e.timestamp for e in browser_events]
-
-    SPATIAL = True
-    if SPATIAL:
-        # Prepare sequences for multidimensional DTW
-        action_sequence = np.array([
-            [e.timestamp, e.mouse_x or 0.0, e.mouse_y or 0.0]
-            for e in action_events
-        ], dtype=np.double)
-
-        browser_sequence = np.array([
-            [e.timestamp, e.message.get('screenX', 0.0), e.message.get('screenY', 0.0)]
-            for e in browser_events
-        ], dtype=np.double)
-
-        # Compute the alignment using multidimensional DTW
-        path = dtw_ndim.warping_path(action_sequence, browser_sequence)
-    else:
-        # Compute the alignment using DTW
-        path = dtw.warping_path(action_timestamps, browser_timestamps)
-
-    # Enforce a one-to-one correspondence by selecting the closest matches
-    filtered_path = enforce_one_to_one_mapping(path, action_timestamps, browser_timestamps)
-
-    match_count = 0
-    mismatch_count = 0
-    error_count = 0  # Initialize error counter
-    remote_time_differences = []  # To store differences in local/remote timestamps for matching events
-    local_time_differences = []  # As above but for local/local
-    mouse_x_differences = []  # To store differences in mouse X positions for matching events
-    mouse_y_differences = []  # To store differences in mouse Y positions for matching events
-
-    logger.info(f"Alignment for {event_type} Events")
-    for i, j in filtered_path:
-        action_event = action_events[i]
-        browser_event = deepcopy(browser_events[j])
-
-        action_event_type = action_event.name.lower()
-        browser_event_type = browser_event.message['eventType'].lower()
-
-        if action_event_type in EVENT_TYPE_MAPPING and browser_event_type == EVENT_TYPE_MAPPING[action_event_type]:
-            match_count += 1
-            remote_time_difference = action_event.timestamp - browser_event.message['timestamp']
-            remote_time_differences.append(remote_time_difference)
-            local_time_difference = action_event.timestamp - browser_event.timestamp
-            local_time_differences.append(local_time_difference)
-
-            # Compute differences between mouse positions
-            if action_event.mouse_x is not None:
-                mouse_x_difference = action_event.mouse_x - browser_event.message['screenX']
-                mouse_y_difference = action_event.mouse_y - browser_event.message['screenY']
-                if mouse_x_difference > 1:
-                    logger.warning(f"{mouse_x_difference=} {action_event.mouse_x=} {browser_event.message['screenX']=}")
-                if mouse_y_difference > 1:
-                    logger.warning(f"{mouse_y_difference=} {action_event.mouse_y=} {browser_event.message['screenY']=}")
-                mouse_x_differences.append(mouse_x_difference)
-                mouse_y_differences.append(mouse_y_difference)
-        else:
-            mismatch_count += 1
-            logger.warning(f"Event type mismatch: Action({action_event_type}) does not match Browser({browser_event_type})")
-
-        logger.info(
-            f"\nAction Event:\n"
-            f"  - Type: {action_event.name}\n"
-            f"  - Timestamp: {action_event.timestamp}\n"
-            f"  - Details: {action_event}\n"
-            f"Browser Event:\n"
-            f"  - Type: {browser_event.message['eventType']}\n"
-            f"  - Timestamp: {browser_event.message['timestamp']}\n"
-            f"  - Details: {browser_event}\n"
-            f"{'-'*80}"
-        )
-
-    logger.info(f"Total Matches: {match_count}")
-    logger.info(f"Total Mismatches: {mismatch_count}")
-
-    # Log unmatched browser events
-    matched_browser_indices = {j for _, j in filtered_path}
-    unmatched_browser_events = [e for idx, e in enumerate(browser_events) if idx not in matched_browser_indices]
-
-    if unmatched_browser_events:
-        logger.warning(f"Unmatched Browser Events: {len(unmatched_browser_events)}")
-        for browser_event in unmatched_browser_events:
-            logger.warning(
-                f"Unmatched Browser Event:\n"
-                f"  - Type: {browser_event.message['eventType']}\n"
-                f"  - Timestamp: {browser_event.message['timestamp']}\n"
-                f"  - Details: {browser_event}\n"
-            )
-            error_count += 1  # Increment error count for each unmatched browser event
-    
-    try:
-        assert len(browser_events) == match_count, "Every BrowserEvent should have a corresponding ActionEvent."
-    except Exception as exc:
-        error_count += 1  # Increment error count for assertion error
-        logger.warning(exc)
-    
-    return error_count, remote_time_differences, local_time_differences, mouse_x_differences, mouse_y_differences
-
-
-def enforce_one_to_one_mapping(path: list, action_timestamps: list, browser_timestamps: list) -> list:
-    """
-    Enforces a one-to-one mapping between BrowserEvents and ActionEvents by selecting the closest match.
-    
-    Args:
-        path: List of tuples representing the DTW path.
-        action_timestamps: List of timestamps for action events.
-        browser_timestamps: List of timestamps for browser events.
-        
-    Returns:
-        filtered_path: List of tuples representing the filtered DTW path with one-to-one mapping.
-    """
-    used_action_indices = set()
-    filtered_path = []
-
-    # Create a dictionary to store the closest match for each browser event
-    closest_matches = {}
-
-    for i, j in path:
-        if j not in closest_matches:
-            closest_matches[j] = (i, abs(action_timestamps[i] - browser_timestamps[j]))
-        else:
-            # If a closer match is found, update the closest match for this browser event
-            current_diff = abs(action_timestamps[i] - browser_timestamps[j])
-            if current_diff < closest_matches[j][1]:
-                closest_matches[j] = (i, current_diff)
-
-    # Collect the closest matches while ensuring each action event is used only once
-    for j, (i, _) in closest_matches.items():
-        if i not in used_action_indices:
-            filtered_path.append((i, j))
-            used_action_indices.add(i)
-
-    return filtered_path
-
-
-if __name__ == "__main__":
-    main()
